requirements.txt
pytube==15.0.0
openai==1.3.6
python-dotenv==1.0.0
